Racket: slightly better handling of the signatures cache
Refreshing the cache on form evaluation (had been deactivated by error
since the introduction of meta-commands). The current behaviour is not
yet completely correct: if one evaluates a form in a modified buffer,
geiser won't notice the new definition's signature, not even after the
buffer is saved if one has gone around asking for autodoc before that.

An improvement would be to make the cache sensitive to file write
times, as suggested back in the day by Eli & Neil.